Split out batched_query into separate query module
